pyproject.toml
[project]
name = "topax"
version = "0.0.1"
dependencies = [
  "numpy",
  "PyOpenGL",
  "PyOpenGL_accelerate",
  "glfw",
  "watchdog",
  "ipython",
  "jinja2"
]

[project.scripts]
topax = "topax._app:main"

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
